requirements.txt
fastapi
uvicorn[standard]
httpx
pandas
openpyxl
python-multipart
xlrd
python-dateutil
